requirements.txt
confluent-kafka==2.2.0
websocket-client==1.6.2
pandas==2.0.3pymongo==4.5.0
pymongo==4.5.0
